generic_cartesian: Do not check start points for safe_dist with DIRECT
Only test the end points of carriages with DIRECT mode for proximity
check and whether their trajectories intersect, but do not test the
start points for proximity. This makes the behavior the same as for
all other modes, and can be useful for carriages homing in the same
direction if homing brings them closer together than safe_distance.

Signed-off-by: Dmitry Butyugin <dmbutyugin@google.com>